# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SCM Ads API documentation. For users with advertiser-scoped roles, JWTs may include an advertiser_ids claim (array). When accessing endpoints with advertiser-scoped permissions, provide X-Advertiser-Id to select the active advertiser scope; if missing, the API may return 400 advertiser_not_selected.",
        "title": "SCM Ads API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/creatives/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Creatives"
                ],
                "summary": "List creatives",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Active advertiser scope (required for advertiser-scoped users)",
                        "name": "X-Advertiser-Id",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Creative"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/creatives/campaign/{campaignID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Creatives"
                ],
                "summary": "List creatives by campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Active advertiser scope (required for advertiser-scoped users)",
                        "name": "X-Advertiser-Id",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Creative"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/creatives/device/{device}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Creatives"
                ],
                "summary": "List creatives by device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device name",
                        "name": "device",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by current day and time",
                        "name": "active_now",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Creative"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/creatives/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Creatives"
                ],
                "summary": "Search creatives",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search text (matches name, type, selected days, time slots, devices)",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/creatives/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Creatives"
                ],
                "summary": "Upload creatives",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Active advertiser scope (required for advertiser-scoped users)",
                        "name": "X-Advertiser-Id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaign_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Selected days (comma separated or repeated)",
                        "name": "selected_days",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Time slots (comma separated or repeated)",
                        "name": "time_slots",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Devices (comma separated or repeated)",
                        "name": "devices",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Creative files",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Creative"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Creative": {
            "type": "object",
            "required": [
                "campaign_id",
                "name",
                "type"
            ],
            "properties": {
                "campaign_id": {
                    "type": "string"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "selected_days": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "time_slots": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "enum": [
                        "image",
                        "video"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CreativeType"
                        }
                    ]
                },
                "uploaded_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.CreativeType": {
            "type": "string",
            "enum": [
                "image",
                "video"
            ],
            "x-enum-varnames": [
                "CreativeTypeImage",
                "CreativeTypeVideo"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token with 'Bearer ' prefix. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}